change application name to BE_SWP, update username, password, change dbname to mydatabase
