+++ src/main/java/com/example/be_swp/Controller/UsersController.java
@@ -1,8 +1,7 @@
-package com.example.demo.Controller;
+package com.example.be_swp.Controller;
 
-import com.example.demo.Models.Users;
-import com.example.demo.Repository.UsersRepository;
-import com.example.demo.Service.UserService;
+import com.example.be_swp.Models.Users;
+import com.example.be_swp.Service.UserService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
