+++ src/main/java/com/example/be_swp/Models/Comments.java
@@ -1,0 +1,4 @@
+package com.example.be_swp.Models;
+
+public class Comments {
+}
