+++ src/main/java/com/example/be_swp/Models/Estheticians_Schedules.java
@@ -1,4 +1,4 @@
-package com.example.demo.Models;
+package com.example.be_swp.Models;
 
 public class Estheticians_Schedules {
 }
