+++ src/main/java/com/example/be_swp/Models/Service.java
@@ -1,0 +1,4 @@
+package com.example.be_swp.Models;
+
+public class Service {
+}
