+++ src/main/java/com/example/be_swp/Repository/RolesRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.Repository;
+package com.example.be_swp.Repository;
 
-import com.example.demo.Models.Roles;
+import com.example.be_swp.Models.Roles;
 import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
