+++ src/main/java/com/example/be_swp/SWPApplication.java
@@ -1,13 +1,13 @@
-package com.example.demo;
+package com.example.be_swp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DemoApplication {
+public class SWPApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(SWPApplication.class, args);
 	}
 
 }
