--- src/main/java/com/example/demo/Models/PaymentDetail.java
@@ -1,4 +1,0 @@
-package com.example.demo.Models;
-
-public class PaymentDetail {
-}
