+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
-spring.application.name=demo
+spring.application.name=BE_SWP
 spring.datasource.username=root
-spring.datasource.password=123
-spring.datasource.url=jdbc:mysql://localhost:3306/springdatadb
+spring.datasource.password=root
+spring.datasource.url=jdbc:mysql://localhost:3306/mydatabase
 
 spring.jpa.generate.dll=true
 spring.jpa.hibernate.ddl-auto=create
