+++ src/test/java/com/example/be_swp/SWPApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.demo;
+package com.example.be_swp;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class SWPApplicationTests {
 
 	@Test
 	void contextLoads() {
